requirements.txt
altgraph==0.17.5
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.59.0
django-recaptcha==4.0.0
et-xmlfile==1.1.0
future==1.0.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
macholib==1.16.4
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.3
pypdf==6.1.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
qrcode==8.2
reportlab==4.4.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.1.0
whitenoise==6.6.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
